🚚 Rename POPUpstream class to POP3Upstream
